Fix call to check_response_size! in sync_read_*
Previously it was checking that the length of the returned array (which is a Vec<Vec<u8>>)
was == size_of(reg_type).

Hilariously this worked for one of my reads, where I was reading 8 values of type u8,
but then it blew up for when I was reading 8 values of type i32.

It's not totally clear to me what value these asserts provide. You'd expect a check inside
read and indeed it appear sthere is one:

```
if status_packet.params().len() != (length as usize * ids.len()) {
            Err(Box::new(CommunicationErrorKind::ParsingError))
        }
```

But for now, I'm changing the check to ensure they byte length of the value
matches the byte length of the type.